order-processing-system/lambdas/create-order/requirements.txt
fastapi==0.109.0
mangum==0.17.0
pydantic[email]==2.5.3
psycopg2-binary==2.9.9
boto3==1.34.14
